flask/requirements.txt
flask-cors==3.0.10
flask-login==0.5.0
flask-migrate==2.5.3
pip-chill==1.0.0
pyjwt==2.0.1
sqlalchemy-serializer==1.3.4.4
